+++ fixtures/StructuredOutput/PolymorphicType/ListItemAge.php
@@ -1,0 +1,24 @@
+<?php
+
+/*
+ * This file is part of the Symfony package.
+ *
+ * (c) Fabien Potencier <fabien@symfony.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Symfony\AI\Fixtures\StructuredOutput\PolymorphicType;
+
+use Symfony\AI\Platform\Contract\JsonSchema\Attribute\With;
+
+final class ListItemAge implements ListItemDiscriminator
+{
+    public function __construct(
+        public int $age,
+        #[With(pattern: '^age$')]
+        public string $type = 'age',
+    ) {
+    }
+}