+++ src/agent/src/StructuredOutput/AgentProcessor.php
@@ -20,9 +20,14 @@ use Symfony\AI\Agent\OutputProcessorInterface;
 use Symfony\AI\Platform\Capability;
 use Symfony\AI\Platform\Result\ObjectResult;
 use Symfony\Component\PropertyInfo\Extractor\PhpDocExtractor;
+use Symfony\Component\PropertyInfo\Extractor\ReflectionExtractor;
 use Symfony\Component\PropertyInfo\PropertyInfoExtractor;
 use Symfony\Component\Serializer\Encoder\JsonEncoder;
+use Symfony\Component\Serializer\Mapping\ClassDiscriminatorFromClassMetadata;
+use Symfony\Component\Serializer\Mapping\Factory\ClassMetadataFactory;
+use Symfony\Component\Serializer\Mapping\Loader\AttributeLoader;
 use Symfony\Component\Serializer\Normalizer\ArrayDenormalizer;
+use Symfony\Component\Serializer\Normalizer\BackedEnumNormalizer;
 use Symfony\Component\Serializer\Normalizer\ObjectNormalizer;
 use Symfony\Component\Serializer\Serializer;
 use Symfony\Component\Serializer\SerializerInterface;
@@ -39,8 +44,20 @@ final class AgentProcessor implements InputProcessorInterface, OutputProcessorIn
         private ?SerializerInterface $serializer = null,
     ) {
         if (null === $this->serializer) {
-            $propertyInfo = new PropertyInfoExtractor([], [new PhpDocExtractor()]);
-            $normalizers = [new ObjectNormalizer(propertyTypeExtractor: $propertyInfo), new ArrayDenormalizer()];
+            $classMetadataFactory = new ClassMetadataFactory(new AttributeLoader());
+            $discriminator = new ClassDiscriminatorFromClassMetadata($classMetadataFactory);
+            $propertyInfo = new PropertyInfoExtractor([], [new PhpDocExtractor(), new ReflectionExtractor()]);
+
+            $normalizers = [
+                new BackedEnumNormalizer(),
+                new ObjectNormalizer(
+                    classMetadataFactory: $classMetadataFactory,
+                    propertyTypeExtractor: $propertyInfo,
+                    classDiscriminatorResolver: $discriminator
+                ),
+                new ArrayDenormalizer(),
+            ];
+
             $this->serializer = new Serializer($normalizers, [new JsonEncoder()]);
         }
     }